Bug 1034143 - Use PContent::OpenAnonymousTemporaryFile for content-process libjar downloads.
The nsIFile subclass is even more fake than RemoteOpenFileChild --
the "native path" is a stringified UUID, because on Unix the file is
unlinked and doesn't have an actual path (and is irrelevant in any case,
because the content process would have no way to open it).

Also, nsDownloader now asserts that it isn't used in a way that would
require directly creating a file from a content process; this is more
polite than a sandbox error, and works on platforms that aren't yet
sandboxed.