Bug 1068838 - Fix file input element tests to open files in the parent process.
The dom/base tests changed here use SpecialPowers.loadChromeScript to
construct a DOM File in the parent and send it to the child, using the
existing remote-blob infrastructure.  The dom/html tests don't need a
real file, so they now construct in-memory files.

As a convenient side-effect, these tests are now enabled for desktop
e10s (they were already being run out-of-process on B2G); most of them
failed before this change due to the code that's being moved/removed.