Bug 1034143 - Use IPC to open temp file for content-process libjar downloads.
This patch uses async IPC to invoke NS_OpenAnonymousTemporaryFile in the
parent and wraps the returned file descriptor in an nsIFile subclass.
The file's "path" is actually a UUID, because this is used only as a
cache key -- on Unix the file is unlinked and is truly anonymous, and
once sandboxing is completed the path won't be useful in any case.

nsJARChannel::AsyncOpen also undergoes some minor rearrangement to deal
with opening the temporary file asynchronously.

Also, nsDownloader now asserts that it isn't used in a way that would
require directly creating a file from a content process; this is more
polite than a low-level sandbox error, and works on platforms that
aren't yet sandboxed enough for that.