Bug 1088387 - Start Linux/B2G sandboxing while the process is single-threaded.
To make this work, any syscalls which would be forbidden by the
seccomp-bpf policy are proxied to an unsandboxed thread, which exits
when the process is ready for sandboxing to be enforced -- i.e., the
process is not actually sandboxed until the thread exits.

This lets us get rid of the inelegant signal-broadcasting code currently
used to enable seccomp-bpf, which will break when pid namespaces are
used.  This also gives us a point to, in future work, manipulate
namespaces and capabilities and have the results apply to the entire
process.