Breakpad Linux client: Terminate process on fake crash signals even if it's pid 1.
A child process can be pid 1 if it's run in a new pid namespace as
part of sandboxing; in this situation the traditional behavior of pid
1 applies and unhandled signals are ignored if possible.  When a crash
signal was sent artificially with kill() or similar, breakpad tries to
resume crashing the process by setting the disposition to SIG_DFL and
reraising it; in this case that doesn't work.

Actual crashes shouldn't have this problem, but it can be useful to
test crash reporting with `kill -11`, and it's very surprising for
the process to respond by reporting a crash but continuing to run --
especially when the application already tried to handle that case by
installing its own signal handler (which Breakpad ignored).

However, Breakpad does _exit() in this case if it fails to signal
itself, so this patch simply extends that to the case where signaling
itself succeeds with no effect.

MozReview-Commit-ID: In7GdWtmYP1