Bug 1439057 - Use the sandbox broker directly for shared memory file creation on Linux. r?gcp r?froydnj
Previously the open() and unlink() calls would raise SIGSYS and each do a
broker RPC, and we had to allow access to all of /dev/shm (allowing the
possibility of interfering with other processes) in the broker policy.

This patch allowing creating an anonymous shared memory file with a
single broker RPC, bypassing the signal handling overhead and avoiding
the need to give access to any other files -- the broker returns the file
already unlinked and the client doesn't specify a filename or path.

The big picture for how this works: base::SharedMemory::Create in a
content process calls into libmozsandbox, which calls a method on its
broker client, which messages the parent process, which calls its own
instance of base::SharedMemory::Create, which doesn't have a broker
client and so does the open/unlink itself.

The broker uses a dedicated thread per child process, so the synchronous
IPC here will interfere only with other brokered file I/O for the same
process, and that was already the case before this patch.  (See also bug
1440203, which will avoid the IPC entirely on newer kernels.)

Bonus fix: the global SandboxReporterClient pointer is made atomic, to
match the new global SandboxBrokerClient pointer.

MozReview-Commit-ID: FuupQjGqy8j