Bug 1400051 - IPC: use process_util_linux on BSD and remove process_util_bsd. r?glandium
This change is mainly to avoid the use of SetAllFDsToCloseOnExec, which
has an unavoidable race condition that can leak file descriptors into
child processes.  The "Linux" implementation of child process creation,
now that B2G support has been removed, is essentially a generic Unix
fork+dup2+execve which works on BSD (and is already used on Solaris).

MozReview-Commit-ID: 71Xa8EVoBqE