Bug 1440203 - Use memfd_create for shared memory on Linux where available. r?froydnj
memfd_create is roughly similar to FreeBSD's SHM_ANON, but it was
introduced in kernel 3.17 and didn't get a C wrapper in glibc until
2.27.  Because we need to be able to build (runtime-detected) support for
it on much older host systems, the syscalls numbers for several common
architectures are hard-coded and syscall() is used directly.

When memfd_create support is detected, the content sandbox policy will
not allow access to /dev/shm; see bug 1439057.

MozReview-Commit-ID: Lv5515Hn8Xq