Bug 1479960 - Get rid of base::SharedMemory::handle. r?froydnj
Despite the comment saying not to use the "handle" except as an opaque
identifier, it is being used to pass the handle to other OS APIs.  Direct
access to the handle needs to be controlled to make sure freezing is
safe, so this patch replaces that with interfaces that are more explicit
about ownership and lifetime.

Differential Revision: https://phabricator.services.mozilla.com/D26740