Bug 1511560 - Allow dup and ftruncate (when needed) in SandboxPolicyCommon. r?gcp
The handling of ftruncate is a little complicated -- it's used for IPC
shared memory, but only when creating segments; so GMP doesn't allow
it and should continue not allowing it, but content needs it and RDD
will as well.

Note that even when we have memfd_create support (bug 1440203),
ftruncate is still necessary even though brokering may not.