Bug 1544825 - Use IPDL type names instead of truncated C++ names in deserialization error messages.  r?nika
Currently, when deserialization fails, the error message contains the base
name of the C++ type, without template parameters; this means we can get
unhelpfully vague errors about `RefPtr` or `Maybe` or `nsTArray`.  (The
identical error cases are then merged by the compiler, so the execution
path can't be recovered from the crash dump.)

This patch uses the IPDL type instead.  It would be possible to invoke
the code generator to use the full C++ type, but more complicated for no
real benefit.