Bug 1294286 - Filter clock IDs in clock_getres sandbox rule. r?gcp
The clockid_t type on Linux has a space of values with encode a pid and
refer to various measures of another process's CPU usage; clock_getres
would, thereby, allow probing whether other processes exist.  This is
a relatively small information leak into the sandboxes, but there's no
reason to allow it.

Differential Revision: https://phabricator.services.mozilla.com/D54081