Bug 1607153: Remove the extra close of ChannelImpl::client_pipe_.
The channel implementation closes the client's end of the socketpair when
it receives input, reasoning that it's safe since the child must have
been launched and inherited its copy and started writing.

However, now that launching is done off the I/O thread, this can race
with the launch thread calling CloseClientFileDescriptor and also trying
to close the socket; this could result in a double-close, and possibly
closing some unrelated fd and causing mysterious breakage somewhere else.

It's also not necessary, because we do call CloseClientFileDescriptor.
So, this patch removes it.