Bug 1615155 - Don't generate default Recv__delete__ if there are args. r?nika
The default method implementations cause problems when trying to
override them with different types in a direct call class.

For the Recv__delete__ case there's a simple solution: omit it if
there are any arguments, because it doesn't make much sense to specify
arguments and then completely ignore them, and the no-arg case isn't a
problem for overriding.