Bug 1470983 - Prelude: Remove use of gtk_widget_get_settings in LookAndFeel.
The GTK nsLookAndFeel implementation uses `gtk_widget_get_settings`
on temporary widgets that aren't attached to anything, which the
documentation says not to do.  Empirically, this seems to return the
same settings object as `gtk_settings_get_default`.

Explicitly using `gtk_settings_get_default` is useful for remote
look-and-feel, so that the parent process can register for property
changes on that object to know when to update the child processes.

Differential Revision: https://phabricator.services.mozilla.com/D94535