Bug 1690921 - Limit IPC `sendmsg` gather list sizes based on socket buffer capacity. r?mccr8
When setting up calls to `sendmsg` for IPC on Unix systems, we generate
`iovec`s for the entire message or until the `IOV_MAX` limit is reached,
whichever comes first.  However, messages can be very large (up to 256
MiB currently), while the OS socket buffer is relatively small (8KiB on
macOS and FreeBSD, ~200KiB on Linux).

This patch detects the socket buffer size with the `SO_SNDBUF` socket
option and cuts off the `iovec` array after it's reached; it also adjusts
the Linux sandbox policy to allow reading that value in all processes.

On Linux the buffer size is complicated; the `socket(7)` man page refers
to doubling the value when setting `SO_SNDBUF` "to allow space for
bookkeeping overhead" and returning that doubled value when it's read.
However, empirically a local-domain socket appears to accept slightly
*more* data (about 3%) than the advertised size.  This is relatively
insignificant; the value from `getsockopt` is used unmodified.

This increases throughput on large messages by about 2.5x on macOS (from
~0.3 to ~0.7 GB/s), but on Linux the improvement is only about 5%.