Bug 1635451 - Minimize content processes' connections to the X server. r?jgilbert,stransky,nika
This patch launches content processes with the `MOZ_HEADLESS` env var set if:
1. the widget toolkit is GTK
2. the display is X11
3. the content process doesn't need GTK for other reasons (e.g., bug 1695381)

The goal is to avoid exhausting Xorg's default limit of 256 clients if
there are many content processes due to Fission.  If these conditions
are met, the content process doesn't need to eagerly connect to the X
server.  This does not affect the sandbox policy, and content processes
can still use X if needed for, e.g.,  WebGL.

The new env var `MOZ_HEADLESS_BROWSER` is added, set for content
processes to indicate if the entire browser is headless (rather than
just a specific process); this is needed in several cases.

The boolean pref `dom.ipc.avoid-x11`, set by default, controls this feature.

Note that disabling `widget.non-native-theme.enabled`, which is also
enabled by default, will restore the use of X11 in all content processes
even if this pref is set; the alternative is that widgets wouldn't render
in that case.

This change should also save us some memory for now-unnecessary
instances of GTK's global state, and may improve content processes
startup time.

Differential Revision: https://phabricator.services.mozilla.com/D112197