Bug 1440103 - Stop leaking the handle returned from CreateProcess in IPC.
Currently, the handle returned from CreateProcess ends up in a closure
in `GeckoChildProcessHost::AsyncLaunch`, where it's used to derive a
handle with more permissions; the original handle is used to resolve the
returned promise, but nothing takes responsibility for closing it.  Note
that this bug predates the use of MozPromise in GeckoChildProcessHost;
previously, the handle was held in a local variable and similarly never
closed.

This patch closes that handle and changes the promise types so
that the process ID is provided instead of a handle; one existing
caller (RemoteSandboxBroker) uses the pid, while others use only the
success/failure bit.