Bug 1658072 - Cleanup: assert that libevent's signal handling code is never used.
Given that libevent's signal handling code is known to have race
conditions, and there are fundamental issues that make it hard to fix
upstream, and previous patches have theoretically removed our last usage
of it, we should assert that it's no longer used.

FIXME: change the message and add a patch to the patch place.