Bug 1746289 - Deprecate the IPDL `intr` keyword and tag legacy uses in tests. r?nika
At this point, the only remaining uses of `intr` are in test files.  To
avoid the possibility of adding any more (until such time as we deal
with the tests and can get rid of `intr` entirely), this patch requires
a `[LegacyIntr]` attribute on any declaration that uses them, and tags
all the uses in tests with that attribute.

Differential Revision: https://phabricator.services.mozilla.com/D133964